Move highlighting tests for pressed keyword to their own file
