+++ tests/test_highlighting/test_if_block.py
@@ -322,65 +322,3 @@ class HighlighterTestIfBlock(HighlightTester):
             "if answer != 246 and answer == 'yes'",
             "KK TTTTTT KK NNN KKK TTTTTT KK SSSSS",
             level=level, lang="en")
-
-    @parameterized.expand([
-        ("level8"),
-        ("level9"),
-        ("level10"),
-        ("level11"),
-        ("level12"),
-        ("level13"),
-        ("level14"),
-        ("level15"),
-        ("level16"),
-        ("level17"),
-    ])
-    def test_if_is_pressed1(self, level):
-        self.assert_highlighted_chr(
-            "if x is pressed",
-            "KK T KK KKKKKKK",
-            level=level, lang="en")
-
-    @parameterized.expand([
-        ("level8"),
-        ("level9"),
-        ("level10"),
-        ("level11"),
-        ("level12"),
-        ("level13"),
-        ("level14"),
-        ("level15"),
-        ("level16"),
-        ("level17"),
-    ])
-    def test_if_is_pressed_print1(self, level):
-        self.assert_highlighted_chr_multi_line(
-            "if x is pressed",
-            "KK T KK KKKKKKK",
-            "    print 'Yay!'",
-            "    KKKKK SSSSSS",
-            level=level, lang="en")
-
-    @parameterized.expand([
-        ("level8"),
-        ("level9"),
-        ("level10"),
-        ("level11"),
-        ("level12"),
-        ("level13"),
-        ("level14"),
-        ("level15"),
-        ("level16"),
-        ("level17"),
-    ])
-    def test_if_is_pressed_print_else_print1(self, level):
-        self.assert_highlighted_chr_multi_line(
-            "if x is pressed",
-            "KK T KK KKKKKKK",
-            "    print 'Yay!'",
-            "    KKKKK SSSSSS",
-            "else",
-            "KKKK",
-            "    print 'Boo!'",
-            "    KKKKK SSSSSS",
-            level=level, lang="en")
